test_apps/requirements.txt
pytest
pytest-embedded-idf
pytest-embedded-serial-esp
pytest-rerunfailures
pytest-timeout